Doing cache Apps.ListRepos
