+++ pkg/gh/jwt.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/whywaita/myshoes/pkg/config"
-	"github.com/whywaita/myshoes/pkg/logger"
 
 	"github.com/google/go-github/v47/github"
 )
@@ -108,32 +107,4 @@ func isInstalledGitHubAppSelected(ctx context.Context, inputScope string, instal
 	default:
 		return fmt.Errorf("%s can't detect scope", inputScope)
 	}
-}
-
-func listAppsInstalledRepo(ctx context.Context, installationID int64) ([]*github.Repository, error) {
-	clientInstallation, err := NewClientInstallation(installationID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create a client installation: %w", err)
-	}
-
-	var opts = &github.ListOptions{
-		Page:    0,
-		PerPage: 100,
-	}
-
-	var repositories []*github.Repository
-	for {
-		logger.Logf(true, "get list of repository from installation, page: %d, now all repositories: %d", opts.Page, len(repositories))
-		lr, resp, err := clientInstallation.Apps.ListRepos(ctx, opts)
-		if err != nil {
-			return nil, fmt.Errorf("failed to get installed repositories: %w", err)
-		}
-		repositories = append(repositories, lr.Repositories...)
-		if resp.NextPage == 0 {
-			break
-		}
-		opts.Page = resp.NextPage
-	}
-
-	return repositories, nil
 }
